Agent Monitor Election
Added new config option to monitors called `maxInstancesPerCluster` that
uses a new generic election process across all agents in a cluster so
that the monitor only runs so many times throughout the entire group of
agents.

This takes the original Kubernetes leader election code that was
hardcoded into the kubernetes-cluster and kubernetes-events monitors and
makes it generic and pluggable with other coordination mechanisms.  A
proof-of-concept memcached implementation has also been added.
Configuration for these pluggable coordinators is done via a new
top-level config option `clustering`.

More than one instance can be requested per cluster, but this requires
that there be at least _n_ instances of the agent to function properly.
This part is still a WIP.