Make optional remote config source values tolerant of source outages
This will allow the agent to be more robust on host startup where
optional config options are dependent on services that might be starting
up along with the agent but should not block the agent from starting.

The optional values will be watched with a 10 second delay which will
cause the value to get populated if present after the service comes up.
This will result in a config refresh in the agent.